Added emerge and pkg to pkg module
Now working on the popup menu with pkg module.
Cannot get pkg rquery %c to display on popup
title. Seems that the popup is appearing before
this happens, probably because the command
is thread run. Need to change to pipe run.

But not tomorrow. Need to work on IMP stuff :-(