Fixed slow local load bug
Problem was with gtk automatically updating iconview with each
change to treemodel. Now we use a backing store for adding
items to treemodel.

We could update partial versions of the treemodel, but instead
we show a status update of the progress in construction
of the treemodel which will be displayed.

Also, localmonitor has now been made public, because when
the backing store tree model changes, the monitor must
point to the active treemodel.