Step 1 towards gtk4
Simple window. Took a while to get WM_CLASS set up so that
i3 can always open Xffm class as floating window. This was
necesary as gtk4 does not have a WM hint for dialog windows.

Seems that this transition will be a bit more challenging than
the move from gtk2 to gtk3.