Finally understood gtk4 css behavior
Instead of using a context style with its own provider for each widget,
a provider is setup for entire application. In the CSS information a
new class is created (with preceeding period), and this class is applied
to the widget which will be customized.

Applied to vertical button box and lpterm output textview so far.