popover sample fully functional
Keeping it simple, we do not use GtkPopoverMenu, as this
apparently requires to complicate things with GIO GMenu
and change entire program algorithm with use of GtkApplication,
which requieres further complications.

Instead, we keep it simple, constructing popup with simple
buttons in a vertical box.

Now we can start putting menus into xffm+/gtk4.