Mask applied to icons of hidden paths
Inverse application of mask, using user selected
background color. Emblems not yet applied...
This in Texture class. Cairo class not used.

Also, removed color submenus and promoted items
to main menu. This because something is broken
in the popover's determination of parent, maybe because
of a race on popdown.

Better to keep it simple with no submenus, but colors
will drop down to a new submenu inside gridview.

	modified:   ../xffm/cairo.hh
	modified:   ../xffm/fm/fm.hh
	modified:   ../xffm/fm/fmbuttonbox.hh
	modified:   ../xffm/fm/window.hh
	modified:   ../xffm/texture.hh
	modified:   ../xffm/util.hh
	modified:   ../xffm/utilbasic.hh
	modified:   ../xffm/workdir.hh
	modified:   ../xffm/xffm.cc
	modified:   ../xffm/xffm.h