Basic navigation functionality running in FileResponse
But code must be cleaned. Currently FileResponsePathbar is included
into FileResponse, but it must exists as a separate class.

Problem is with reload_f(), which acts on FileResponse, but
must be configured in FileResponsePathbar.

Should just add a pointer to reload_f and configure that on constructor,
just as done with path_.

Time to go now...