First step in backward communication between dialogs
Now the mkdir dialog in EFS response will set the label in
the parent FileResponse dialog.

Cool. Now we need to set the label on selection, initial
selection and moving selection.

Also, rehabilitate the mkdir command.

Once that is done, then use the same class template schema
to pass on the FileSelection to the EFS corresponding entry.

That would make EFS almost ready.