Eliminate the TimeoutResponse class
Code is still there, but comented out. Apparently gtk4 does not
do too well with switching to main context thread. It seems as
though gtk4 returns to gtk1 behaviour in which more than one
thread can perform gtk, and handles things internally.
This is what seems apparent. The problem with the timeout
thread is that the thread is trying to call the window destroy.

Maybe if the thread just does a call to unref things will work
fine. But anyways, the timeout routine is made obsolete by the
new main window dialogs, so, so long...