Reenable gtk_notebook_set_scrollable for testing
This is at home in a i5 4core laptop.
Error does not occur. Did not check with
valgrind so I do not know if invalid write
is gone. Will check tomorroe.

Maybe the invalid write is gone with the
rubber band disabling.

Anyways, in the end I no longer fancy
using the notebook_scrollable, with
or without bug. But for now we are
in testing mode.

	modificados:     ../xffm/fm/classes/ecryptfs.hh
	modificados:     ../xffm/fm/templates/window.hh
	modificados:     ../xffm/response/templates/efsresponse.hh