Icon size enhancements.
New icon size limit, 768.
Iconsize set on a path basis. When not set by
user, defaults to 48.
Bookmarks, fstab views currently not applicable (just
as with regexp filtering).

This is useful like for setting a regexp to "png$"
to only show the png files and zooming up to a
good viewing size.

	modificados:     ../xffm/TODO
	modificados:     ../xffm/fm/classes/fmbuttonbox.hh
	modificados:     ../xffm/fm/templates/factory.hh
	modificados:     ../xffm/fm/templates/window.hh
	modificados:     ../xffm/response/templates/regexpresponse.hh