Remove item from TODO
Apparently gtk4 does not copy entry text to primary clipboard
when text is selected, but does do the copy except when it is
selected by double click on empty space of entry.

This is a change (perhaps) in gtk behaviour and is precisely
what I was thinking about doing so that selection is copied
to primary clipboard. No longer necessary.

	modificados:     ../xffm/TODO